Changing pointers to 8Byte types to compile with 64-bit compilers.
git-svn-id: file:///home/chulwoo/CPS/repo/CVS2SVN/cps/trunk@2492 fb734a29-28db-4bc5-88a3-3cb5ee595545

+++ cps_pp/include/comms/nga_reg.h
@@ -3,13 +3,13 @@ CPS_START_NAMESPACE
 //--------------------------------------------------------------------
 //  CVS keywords
 //
-//  $Author: zs $
-//  $Date: 2004-08-18 11:57:36 $
-//  $Header: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/include/comms/nga_reg.h,v 1.4 2004-08-18 11:57:36 zs Exp $
-//  $Id: nga_reg.h,v 1.4 2004-08-18 11:57:36 zs Exp $
+//  $Author: chulwoo $
+//  $Date: 2006-09-18 05:07:39 $
+//  $Header: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/include/comms/nga_reg.h,v 1.5 2006-09-18 05:07:39 chulwoo Exp $
+//  $Id: nga_reg.h,v 1.5 2006-09-18 05:07:39 chulwoo Exp $
 //  $Name: not supported by cvs2svn $
 //  $Locker:  $
-//  $Revision: 1.4 $
+//  $Revision: 1.5 $
 //  $Source: /home/chulwoo/CPS/repo/CVS/cps_only/cps_pp/include/comms/nga_reg.h,v $
 //  $State: Exp $
 //
@@ -47,7 +47,7 @@ enum{ CRAM_SCRATCH_ADDR = 0x809F95 - CRAM_SCRATCH_SIZE };
 #else
 // initialized in ../cbuf_sim/cram.C where memory is allocated
 // in order to imitate the scratch cram.
-extern unsigned int CRAM_SCRATCH_ADDR ; 
+extern unsigned long CRAM_SCRATCH_ADDR ; 
 #endif
 
 #endif
